backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
email-validator==2.1.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.8
itsdangerous==2.2.0

# Queue
redis[hiredis]==5.0.4
celery[redis]==5.4.0
flower==2.0.1

# AI/NLP (all Apache-2.0/MIT)
spacy==3.7.4
transformers==4.41.2
torch==2.3.0
sentence-transformers==3.0.0
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
langdetect==1.0.9

# Crawling
playwright==1.44.0
httpx==0.27.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2

# OSS replacements
aiosmtplib==3.0.1        # replaces SendGrid
minio==7.2.7             # replaces AWS S3
hvac==2.1.0              # replaces AWS Secrets Manager (Vault)
jinja2==3.1.4

# Observability
prometheus-client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
sentry-sdk[fastapi]==2.5.1  # wire-compatible with GlitchTip OSS
structlog==24.2.0
slowapi==0.1.9

# Dev/Test/Security
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
factory-boy==3.3.0
faker==25.3.0
black==24.4.2
ruff==0.4.7
mypy==1.10.0
bandit==1.7.8
safety==3.2.3
pip-audit==2.7.3
